+++ src/main/java/dev/lambdaurora/lambdabettergrass/LambdaBetterGrass.java
@@ -27,11 +27,12 @@ import org.jetbrains.annotations.NotNull;
  * Represents the LambdaBetterGrass mod.
  *
  * @author LambdAurora
- * @version 1.2.1
+ * @version 1.2.2
  * @since 1.0.0
  */
 public class LambdaBetterGrass implements ClientModInitializer {
 	public static final String NAMESPACE = "lambdabettergrass";
+	@Deprecated(forRemoval = true) // JsonParser constructor is deprecated in 1.18+
 	public static final JsonParser JSON_PARSER = new JsonParser();
 	/* Default masks */
 	public static final Identifier BETTER_GRASS_SIDE_CONNECT_MASK = mc("bettergrass/mask/standard_block_side_connect.png");