+++ src/main/java/dev/lambdaurora/lambdabettergrass/util/LayeredBlockUtils.java
@@ -9,45 +9,36 @@
 
 package dev.lambdaurora.lambdabettergrass.util;
 
-import dev.lambdaurora.lambdabettergrass.LambdaBetterGrass;
 import dev.lambdaurora.lambdabettergrass.metadata.LBGLayerState;
 import dev.lambdaurora.lambdabettergrass.metadata.LBGState;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
-import net.minecraft.client.render.model.UnbakedModel;
 import net.minecraft.state.property.Property;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.BlockRenderView;
-import org.jetbrains.annotations.Nullable;
 
-import java.util.function.Function;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Represents utilities about snow.
  *
  * @author LambdAurora
- * @version 1.2.1
+ * @version 1.2.2
  * @since 1.0.0
  */
 public final class LayeredBlockUtils {
-	private static final Direction[] DIRECTIONS = Direction.values();
+	private static final List<Direction> HORIZONTAL_DIRECTIONS = Arrays.stream(Direction.values())
+			.filter(dir -> dir.getAxis().isHorizontal())
+			.collect(Collectors.toList());
 
 	private LayeredBlockUtils() {
 		throw new UnsupportedOperationException("LayeredBlockUtils only contains static definitions.");
-	}
-
-	/**
-	 * Returns the unbaked model of snow layer.
-	 *
-	 * @param modelGetter the model getter
-	 * @return the unbaked model
-	 */
-	public static @Nullable UnbakedModel getSnowLayerModel(Function<Identifier, UnbakedModel> modelGetter) {
-		return modelGetter.apply(LambdaBetterGrass.mc("block/snowy_layer"));
 	}
 
 	public static boolean shouldGrassBeSnowy(BlockRenderView world, BlockPos pos, Identifier stateId, BlockState upState, boolean onlyPureSnow) {
@@ -92,32 +83,24 @@ public final class LayeredBlockUtils {
 
 	public static int getNearbyLayeredBlocks(BlockRenderView world, BlockPos pos, Block layerBlock, Block type, boolean onlySourceBlock) {
 		int nearbySnow = 0;
-		for (var direction : DIRECTIONS) {
-			if (direction.getAxis().isHorizontal()) {
-				var offsetPos = pos.offset(direction);
-				var block = world.getBlockState(offsetPos).getBlock();
-				if (block == type && !onlySourceBlock) {
-					if (getNearbySnowLayers(world, offsetPos) > 1)
-						nearbySnow++;
-				} else if (block == layerBlock) {
+		for (var direction : HORIZONTAL_DIRECTIONS) {
+			var offsetPos = pos.offset(direction);
+			var block = world.getBlockState(offsetPos).getBlock();
+			if (block == type && !onlySourceBlock) {
+				if (getNearbyBlockLayers(world, offsetPos, layerBlock) > 1)
 					nearbySnow++;
-				}
+			} else if (block == layerBlock) {
+				nearbySnow++;
 			}
 		}
 		return nearbySnow;
 	}
 
-	public static int getNearbySnowLayers(BlockRenderView world, BlockPos pos) {
-		return getNearbyBlockLayers(world, pos, Blocks.SNOW);
-	}
-
 	public static int getNearbyBlockLayers(BlockRenderView world, BlockPos pos, Block layerBlock) {
 		int nearbySnow = 0;
-		for (var direction : DIRECTIONS) {
-			if (direction.getAxis().isHorizontal()) {
-				if (world.getBlockState(pos.offset(direction)).getBlock() == layerBlock)
-					nearbySnow++;
-			}
+		for (var direction : HORIZONTAL_DIRECTIONS) {
+			if (world.getBlockState(pos.offset(direction)).getBlock() == layerBlock)
+				nearbySnow++;
 		}
 		return nearbySnow;
 	}